Start changing code to allow for game saving
In this commit I changed the play class to allow you to initialize a
game that had already started. I added some code to the output class to
be printed if the user is starting a game in the middle. I also cleaned
up a few other parts of the code here and there.

The next step will be to create a way to export the game in JSON format
and import it again.