Polish code and make room for more cases
This commit polished up the code and made the game completely play-able.
Previously the modules were being treated like classes still, so this
commit shifted them over to full modules where they were included in the
Play class. The saving module was edited to delete a file after it was
loaded so that the games folder doesn't get full of already-played games
. Other small code issues were fixed.